apps/service-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-api"
version = "1.0.0"
description = "FastAPI backend for portfolio with AI chat, RAG, and analytics"
authors = [{name = "Surya", email = "surya@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "starlette==0.27.0",
    "uvicorn[standard]==0.24.0",
    "httpx==0.25.2",
    "pydantic==2.5.0",
    "numpy==1.24.3",
    "python-frontmatter==1.0.0",
    "markdown==3.5.1",
    "scikit-learn==1.3.2",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["data*", "scripts*", "tests*"]

[tool.ruff]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long
line-length = 100
target-version = "py311"

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "scripts/*"]